fix bugs

+++ services/moodLogs_service.go
@@ -5,6 +5,7 @@ import (
 	models "moodly/Models"
 	"moodly/repositories"
 	"strings"
+	"time"
 )
 
 type MoodLogsService struct {
@@ -23,8 +24,12 @@ func (s *MoodLogsService) CreateMoodLog(moodLog *models.MoodLog) error {
 		return errors.New("user id is required")
 	}
 
-	if moodLog.Mood == 0 {
-		return errors.New("mood is required")
+	if err := validateMood(moodLog.Mood); err != nil {
+		return err
+	}
+
+	if moodLog.Causes == "" {
+		return errors.New("causes is required")
 	}
 
 	return s.repo.CreateMoodLog(moodLog)
@@ -41,10 +46,18 @@ func (s *MoodLogsService) GetMoodLogsByDate(userID uint, date string) ([]models.
 		return nil, errors.New("date is required")
 	}
 
-	return s.repo.FindMoodLogsByDate(userID, date)
+	parsedDate, err := time.Parse("2006-01-02", date)
+	if err != nil {
+		return nil, errors.New("invalid date format")
+	}
+
+	return s.repo.FindMoodLogsByDate(userID, parsedDate)
 }
 
 func (s *MoodLogsService) UpdateMoodLog(moodLog *models.MoodLog) error {
+	moodLog.Note = strings.TrimSpace(moodLog.Note)
+	moodLog.Causes = strings.TrimSpace(moodLog.Causes)
+
 	if moodLog.ID == 0 {
 		return errors.New("mood log id is required")
 	}
@@ -53,8 +66,12 @@ func (s *MoodLogsService) UpdateMoodLog(moodLog *models.MoodLog) error {
 		return errors.New("user id is required")
 	}
 
-	if moodLog.Mood == 0 {
-		return errors.New("mood is required")
+	if err := validateMood(moodLog.Mood); err != nil {
+		return err
+	}
+
+	if moodLog.Causes == "" {
+		return errors.New("causes is required")
 	}
 
 	return s.repo.UpdateMoodLog(moodLog)
@@ -65,5 +82,17 @@ func (s *MoodLogsService) DeleteMoodLog(id uint, userID uint) error {
 		return errors.New("mood log id is required")
 	}
 
+	if userID == 0 {
+		return errors.New("user id is required")
+	}
+
 	return s.repo.DeleteMoodLog(id, userID)
+}
+
+func validateMood(mood int) error {
+	if mood < 1 || mood > 5 {
+		return errors.New("mood must be between 1 and 5")
+	}
+
+	return nil
 }
